Rename test files for Django 1.6+ test discovery.
