+++ rtwilio/tests/__init__.py
@@ -1,3 +1,3 @@
-from rtwilio.tests.forms import *
-from rtwilio.tests.outgoing import *
-from rtwilio.tests.views import *
+from .test_forms import *
+from .test_outgoing import *
+from .test_views import *
